fix(macos): stabilize required test gate
