+++ macos/CrabfleetMac/Tests/CrabfleetMacTests/QUICTransportTests.swift
@@ -82,8 +82,8 @@ struct QUICTransportTests {
     defer { scope.remove() }
 
     let created = try scope.loadOrCreate()
-    let certificate = try #require(certificate(of: created.identity))
-    let publicKeyData = try #require(publicKeyExternalRepresentation(of: certificate))
+    let hostCertificate = try #require(certificate(of: created.identity))
+    let publicKeyData = try #require(publicKeyExternalRepresentation(of: hostCertificate))
     let privateKey = try #require(storedPrivateKey(applicationTag: scope.applicationTag))
 
     // Reproduce the pre-fix leak: extra distinct certificates (random serials)
@@ -153,6 +153,7 @@ struct QUICTransportTests {
     .timeLimit(.minutes(1)))
   @MainActor
   func tailnetQUICListenerAuthenticatesWithARD() async throws {
+    try await RFBARDPrewarmer.shared.prepare()
     let fixture = try QUICIdentityFixture()
     defer { fixture.remove() }
     let tcpPort = try availableLoopbackPort(socketType: SOCK_STREAM)
@@ -452,6 +453,7 @@ struct QUICTransportTests {
 
   @Test(.enabled(if: udpLoopbackAvailable, "UDP loopback is unavailable in this sandbox"))
   func keepsTCPListenerReadyWhenQUICPortIsOccupied() async throws {
+    try await RFBARDPrewarmer.shared.prepare()
     let fixture = try QUICIdentityFixture()
     defer { fixture.remove() }
     let blocker = try UDPPortReservation()
@@ -499,6 +501,7 @@ struct QUICTransportTests {
     .enabled(if: udpLoopbackAvailable, "UDP loopback is unavailable in this sandbox"),
     .timeLimit(.minutes(1)))
   func connectionGroupsWithoutStreamsDoNotReserveViewerSlots() async throws {
+    try await RFBARDPrewarmer.shared.prepare()
     let fixture = try QUICIdentityFixture()
     defer { fixture.remove() }
     let tcpPort = try availableLoopbackPort(socketType: SOCK_STREAM)
@@ -578,30 +581,6 @@ struct QUICTransportTests {
     #expect(gate.activeCount == 0)
     #expect(gate.reservedCount == 0)
   }
-}
-
-private func availableLoopbackPort(socketType: Int32) throws -> UInt16 {
-  let descriptor = socket(AF_INET, socketType, 0)
-  guard descriptor >= 0 else { throw POSIXError(.ENOTSOCK) }
-  defer { close(descriptor) }
-  var address = sockaddr_in()
-  address.sin_len = UInt8(MemoryLayout<sockaddr_in>.size)
-  address.sin_family = sa_family_t(AF_INET)
-  address.sin_addr = in_addr(s_addr: inet_addr("127.0.0.1"))
-  let bound = withUnsafePointer(to: &address) {
-    $0.withMemoryRebound(to: sockaddr.self, capacity: 1) {
-      bind(descriptor, $0, socklen_t(MemoryLayout<sockaddr_in>.size)) == 0
-    }
-  }
-  guard bound else { throw POSIXError(POSIXErrorCode(rawValue: errno) ?? .EINVAL) }
-  var length = socklen_t(MemoryLayout<sockaddr_in>.size)
-  let resolved = withUnsafeMutablePointer(to: &address) {
-    $0.withMemoryRebound(to: sockaddr.self, capacity: 1) {
-      getsockname(descriptor, $0, &length) == 0
-    }
-  }
-  guard resolved else { throw POSIXError(POSIXErrorCode(rawValue: errno) ?? .EINVAL) }
-  return UInt16(bigEndian: address.sin_port)
 }
 
 private final class UDPPortReservation {
